+++ pkg/analyzer/lib/dart/ast/ast.dart
@@ -916,6 +916,10 @@ abstract class ClassTypeAlias implements TypeAlias {
   SimpleIdentifier get name;
 
   /// Return the name of the superclass of the class being declared.
+  NamedType get superclass;
+
+  /// Return the name of the superclass of the class being declared.
+  @Deprecated('Use superclass instead')
   NamedType get superclass2;
 
   /// Return the type parameters for the class, or `null` if the class does not
@@ -1342,6 +1346,10 @@ abstract class ConstructorName implements AstNode, ConstructorReferenceNode {
   Token? get period;
 
   /// Return the name of the type defining the constructor.
+  NamedType get type;
+
+  /// Return the name of the type defining the constructor.
+  @Deprecated('Use type instead')
   NamedType get type2;
 }
 
@@ -1783,6 +1791,10 @@ abstract class ExtendsClause implements AstNode {
   Token get extendsKeyword;
 
   /// Return the name of the class that is being extended.
+  NamedType get superclass;
+
+  /// Return the name of the class that is being extended.
+  @Deprecated('Use superclass instead')
   NamedType get superclass2;
 }
 
@@ -2680,6 +2692,10 @@ abstract class ImplementsClause implements AstNode {
   Token get implementsKeyword;
 
   /// Return the list of the interfaces that are being implemented.
+  NodeList<NamedType> get interfaces;
+
+  /// Return the list of the interfaces that are being implemented.
+  @Deprecated('Use interfaces instead')
   NodeList<NamedType> get interfaces2;
 }
 
@@ -3532,6 +3548,10 @@ abstract class OnClause implements AstNode {
   Token get onKeyword;
 
   /// Return the list of the classes are superclass constraints for the mixin.
+  NodeList<NamedType> get superclassConstraints;
+
+  /// Return the list of the classes are superclass constraints for the mixin.
+  @Deprecated('Use superclassConstraints instead')
   NodeList<NamedType> get superclassConstraints2;
 }
 
@@ -4615,6 +4635,10 @@ abstract class WhileStatement implements Statement {
 /// Clients may not extend, implement or mix-in this class.
 abstract class WithClause implements AstNode {
   /// Return the names of the mixins that were specified.
+  NodeList<NamedType> get mixinTypes;
+
+  /// Return the names of the mixins that were specified.
+  @Deprecated('Use mixinTypes instead')
   NodeList<NamedType> get mixinTypes2;
 
   /// Return the token representing the 'with' keyword.