+++ pkg/analyzer/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added `getField` to `ExtensionElement`.
 * Added `isGenerative` to `ConstructorElement`.
 * Added `isDartCoreEnum` to `ClassElement` and `DartType`.
+* Deprecated `superclass2`, `mixinTypes2`, etc. Use `superclass`, `mixinTypes`, etc.
 
 ## 3.2.0
 * Deprecated `changes` getter in `File` and `Folder`, use `watch()` instead.